requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
async-timeout
asyncpg
attrs
certifi
click
deprecation
dnspython
email_validator
exceptiongroup
fastapi
fastapi-cli
frozenlist
gotrue
h11
h2
hpack
httpcore
httptools
httpx
hyperframe
idna
iniconfig
Jinja2
markdown-it-py
MarkupSafe
mdurl
multidict
packaging
pluggy
postgrest
propcache
pydantic
pydantic_core
Pygments
PyJWT
pytest
pytest-mock
python-dateutil
python-dotenv
python-multipart
PyYAML
realtime
rich
rich-toolkit
shellingham
six
sniffio
starlette
storage3
StrEnum
supabase
supafunc
tomli
typer
typing-inspection
typing_extensions
uvicorn
uvloop
watchdog
watchfiles
websockets
yarl
